vswarm-like/model_serving/cnn_image_classification/requirements.txt
grpcio==1.45.0
grpcio-tools==1.45.0
opentelemetry-api==1.3.0
opentelemetry-exporter-zipkin==1.3.0
opentelemetry-exporter-zipkin-json==1.3.0
opentelemetry-exporter-zipkin-proto-http==1.3.0
opentelemetry-instrumentation==0.22b0
opentelemetry-instrumentation-grpc==0.22b0
opentelemetry-sdk==1.3.0
opentelemetry-semantic-conventions==0.22b0
grpcio_reflection
tensorflow
keras
numpy
